% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NASAPOWER.R
\name{climdata}
\alias{climdata}
\title{Download NASA POWER data}
\usage{
climdata(t = NULL, lon = NULL, lat = NULL, loc.id = NULL, ...)
}
\arguments{
\item{t}{Is a dataframe with longitud and latitude in decimal formats and a location ID.}

\item{lon}{Is the column index in t that contains the longitud in decimal format}

\item{lat}{Is the column index in t that contains the latitude in decimal format}

\item{loc.id}{Is column name of the location ids in the dataframe t}

\item{...}{Arguments passed to the get_power function}
}
\description{
Downloads climate data from the NASA POWER data base. It is useful for situations when data is required from seeveral sites.
}
