% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords format.R
\name{format.coords}
\alias{format.coords}
\title{Changes the coordinates to decimal degrees from IWIS raw site files. It adds two columns to the input dataframe: lat and lon}
\usage{
\method{format}{coords}(coords)
}
\arguments{
\item{coords}{a data frame from IWIS database that contains the columns Long_degrees, Long_minutes, Lat_degrees and Lat_minutes.}
}
\description{
Changes the coordinates to decimal degrees from IWIS raw site files. It adds two columns to the input dataframe: lat and lon
}
\examples{
coords = format.coords(coords)
}
