% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G matrix functions.R
\name{G.matrix}
\alias{G.matrix}
\title{Function to generate the G matrix from marker data.}
\usage{
G.matrix(X = x, method = "cross.product.average")
}
\arguments{
\item{X}{Is a Marker matrix with marker Ids are in columns and Individuals are in rows}

\item{method}{One of  "cross.product", "cross.product.average" or "VanRaden". The default methods is "cross.product.average".
"cross.product is simply XX'.
"cross.product.average" is X centered and scaled and then XX' divided the number of markers
"VanRaden" is X centered and scaled by the allele frequency as in VanRaden 2007.}
}
\description{
Function to generate the G matrix from marker data.
}
\examples{
G = G.matrix(X = x, method = "VanRaden")
}
