% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H2.R
\name{h2}
\alias{h2}
\title{Broad sense heritability in G + E + GxE models}
\usage{
h2(lm = lm, varg = "g", varge = "g:trt", reps = NULL, envs.col = NULL)
}
\arguments{
\item{lm}{Is an lme4 model object with a GxE term.}

\item{varg}{Is the genetic variance as expressed in the lme4 model object}

\item{varge}{is the GxE variance as expressed in the lme4 object}

\item{reps}{Is the name of the column in the data frame that contains the reps}

\item{envs.col}{Is the name of the column in the data frame that contains the envrionments, treatments, etc.}
}
\description{
h2 is a function to estimate broad sense heritaiblity of multienvironmental trials from the variance components of an lme4 model object.
It only considers the two-way interaction between genotypes and environments.
}
