% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H2 functions.R
\name{h2.single}
\alias{h2.single}
\title{h2.single Isa function to calculate the broad sense heritability from an lme4 model object that does not contain GxE interaction
It only considers the two-way interaction between genotypes and environments.}
\usage{
h2.single(lm = lm, varg = "g", reps = NULL)
}
\arguments{
\item{lm}{Is an lme4 model object with a GxE term.}

\item{reps}{Is the name of the column in the data frame that contains the reps}
}
\description{
h2.single Isa function to calculate the broad sense heritability from an lme4 model object that does not contain GxE interaction
It only considers the two-way interaction between genotypes and environments.
}
\examples{
H2 = h2.single(lm = lm, varg = "g", reps = "reps")
}
