% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H2.R
\name{h2single}
\alias{h2single}
\title{Broad sense heritability in one way models}
\usage{
h2single(lm = lm, varg = "g", reps = NULL)
}
\arguments{
\item{lm}{Is an lme4 model object with a GxE term.}

\item{varg}{Is the genetic variance as expressed in the lme4 model object}

\item{reps}{Is the name of the column in the data frame that contains the reps}
}
\description{
h2.single Isa function to calculate the broad sense heritability from an lme4 model object that does not contain GxE interaction
It only considers the two-way interaction between genotypes and environments.
}
