% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp format functions.R
\name{hap.to.doubl}
\alias{hap.to.doubl}
\title{Function to convert HAPMAP format to double alleles with no separation, i.e, from A to AA, C to CC, R to AG, Y to CT, etc.,}
\usage{
hap.to.doubl(df = df, snp.col = NULL, is.gid.in.col = FALSE, snp.names = NULL)
}
\arguments{
\item{df}{Is a data frame that contains SNP data in IUPAC format}

\item{snp.col}{Is the index where the SNP data starts in the data frame. Set to NULL if the SNPs are in rows}

\item{is.gid.in.col}{Is a Logical to indicate if the individuals are in columns (TRUE) or in rows (FALSE). Default value is FALSE}

\item{snp.names}{Is an index of the column where the SNP ids are if is.gid.in.col = TRUE}
}
\description{
Function to convert HAPMAP format to double alleles with no separation, i.e, from A to AA, C to CC, R to AG, Y to CT, etc.,
}
