% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HapmapToNumeric.R
\name{hapmaptonumeric}
\alias{hapmaptonumeric}
\title{Recodes single nucleotides to numeric format}
\usage{
hapmaptonumeric(gd.snp)
}
\arguments{
\item{gd.snp}{A single SNP vector in hapmap format}
}
\value{
Returns a SNP vector in numeric format
}
\description{
This function takes one SNP marker with the single nucleotide hapmap format of snps to convert them to numeric format:
0, 1, 2. Where 0 and 2 are the homozygous genotypes and 1 the herozygote genotypes. 0 is assigned to the major allele
and 2 to the minor allele.
}
