% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp format functions.R
\name{hpmtonumericadd}
\alias{hpmtonumericadd}
\title{Hapmap to numeric}
\usage{
hpmtonumericadd(df = d, g.names = NULL, snp.col = NULL)
}
\arguments{
\item{df}{Is a dataframe where the markers are in columns and the gids are in rows}

\item{g.names}{Is a vector of genotype ids (gids) taken from df}

\item{snp.col}{Is the index where the marker data starts in the df}
}
\description{
This function converts the hapmap format to numeric by using the SNPassoc package.
}
