% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp format functions.R
\name{rm.minor.allele}
\alias{rm.minor.allele}
\title{This function removes markers at certain minor allele freq treshold}
\usage{
rm.minor.allele(X = x, minor.threshold = 0.05, plot.minor.allele = F)
}
\arguments{
\item{X}{Is a marker matrix with markers in columns and gids in rows}

\item{minor.threshold}{Is a threshold for the minor allele frequency (MAF)}

\item{plot.minor.allele}{Logial to indicate if the histogram of the MAF is displayed}
}
\description{
This function removes markers at certain minor allele freq treshold
}
