% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp format functions.R
\name{rmhetsnp}
\alias{rmhetsnp}
\title{Remove heterozygotes in X}
\usage{
rmhetsnp(X = NULL, threshold = 10, plot.het = F)
}
\arguments{
\item{X}{Is a marker matrix with markers in columns and gids in rows}

\item{threshold}{Is a threshold for the \% of the heterozygotes in X}

\item{plot.het}{Logical to indicate if the histogram of heterozygocity is displayed. Default is FALSE}
}
\description{
This function is to remove the markers that have a determined proportion of heterozygotes
}
