% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RemovingMissngGenotypes.R
\name{rmMissingGenotypes}
\alias{rmMissingGenotypes}
\title{Removing missing genotypes/lines/entries/GIDs}
\usage{
rmMissingGenotypes(X = NULL, threshold = 50, plot.missing = FALSE)
}
\arguments{
\item{X}{A marker matrix, usually obtained from the routines in FPGC}

\item{threshold}{A threshold of missing marker information for genotypes in percentage}

\item{plot.missing}{Logical. Indicating if a histograms of missing information printed}
}
\description{
This function removes genotyped individuals from a marker matrix that do not meet
the minimum threshold of marker information
}
