% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snptonumeric.R
\name{snpMatrixToNumeric}
\alias{snpMatrixToNumeric}
\title{snpMatrixToNumeric}
\usage{
snpMatrixToNumeric(X = NULL)
}
\arguments{
\item{X}{A SNP matrix containing in its row names the id of the individuals and columns as SNPs}
}
\value{
A numeric SNP matrix
}
\description{
Recodes a biallelic SNP matrix in hapmap format to numeric format, e.g., 0, 1, 2.
}
