% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snptonumeric.R
\name{snptonumeric}
\alias{snptonumeric}
\title{Recodes double nucleotides to numeric format}
\usage{
snptonumeric(snp = NULL)
}
\arguments{
\item{snp}{A single SNP vector or dataframe in hapmap format, e.g., two letters for nucleotides without using characters for separation.}
}
\value{
Returns a SNP vector in numeric format
}
\description{
Recodes a single snp in hapmap format (double nubleotide letters i.e: AA, CC, GG, etc) to numeric format.
For biallelic SNPs, homozygous are 0 and 2, and heterozygotes are 1.
}
