% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trials stats for lists.R
\name{trial.stats.list}
\alias{trial.stats.list}
\title{Function suited to prcess lme4 model objects to calculate the broad sense heritability, and statistical parameter of field experiments,
i.e, CV, ASED, LSD, etc.,}
\usage{
trial.stats.list(
  lm = lm,
  varg = "g",
  varge = NULL,
  reps = NULL,
  envs.col = NULL,
  means.gxe = F,
  which.means = NULL,
  ASED = F,
  gxe.model = F,
  save.output = T,
  file.name = "line-means"
)
}
\arguments{
\item{lm}{Is an lme4 model object. It is a model where the genotypes are included as random effect to extract the G variance}

\item{varg}{Is the genetic variance as expressed in the lme4 model object}

\item{varge}{is the GxE variance as expressed in the lme4 object}

\item{reps}{Is the name of the column in the data frame that contains the reps}

\item{envs.col}{Is the name of the column in the data frame that contains the envrionments, treatments, etc.}

\item{means.gxe}{Logical to indicate if the GxE adjusted means will be estimated}

\item{which.means}{A character to indicate which adusted means will be extracted from the model.}

\item{ASED}{Logical to indicate if the ASED will be calculated. Default is FALSE, to speed computation since in this implementation ASED requires
estimation of all pairwise comparisons of means which can be slow if the data frame is large}

\item{gxe.model}{Logical to indicate if the lme4 object model is a GxE type}

\item{save.output}{Logical to indicate if the outuput of adjuste means and trial stats will be saved in the computer}

\item{file.name}{Character to indicate the name of the file if save.output = TRUE.}
}
\description{
Function suited to prcess lme4 model objects to calculate the broad sense heritability, and statistical parameter of field experiments,
i.e, CV, ASED, LSD, etc.,
}
