% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extract VarComps lme4.R
\name{var.comps}
\alias{var.comps}
\title{Extraction of variance components from an lme4 object}
\usage{
var.comps(lm)
}
\arguments{
\item{lm}{Is a lme4 model objeject
#' @examples
vc = var.comps(lm)}
}
\description{
Extraction of variance components from an lme4 object
}
