% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractVarCompsLme4.R
\name{varcomps}
\alias{varcomps}
\title{Extract variance components}
\usage{
varcomps(lm)
}
\arguments{
\item{lm}{Is a lme4 model objeject}
}
\description{
Extraction of variance components from an lme4 object
}
