% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp format functions.R
\name{XImp}
\alias{XImp}
\title{Cleans Impute}
\usage{
XImp(X, out = out)
}
\arguments{
\item{X}{the output of the Impute function}

\item{out}{is the out element in the list from the Impute function}
}
\description{
This function takes the output of the Impute function to keep the adequate data matrix. Run after Impute
}
