% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NASAPOWER.R
\name{climdata}
\alias{climdata}
\title{Download NASA POWER data}
\usage{
climdata(t = NULL, lon = NULL, lat = NULL, new.names = NULL, ...)
}
\arguments{
\item{t}{Is a dataframe with longitud and latitude in decimal formats and a "Loc_no" column as location ID.}

\item{lon}{Is the column index in t that contains the longitud in decimal format}

\item{lat}{Is the column index in t that contains the latitude in decimal format}

\item{new.names}{Is a vector of new names for the variables to be extracted. Used for simplicity in naming the variables.}

\item{...}{Arguments passed to the get_power function}
}
\description{
Downloads climate data from the NASA POWER data base. This functions requires the nasapower library to be installed
}
