% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNPformat.R
\name{Impute}
\alias{Impute}
\title{Impute X}
\usage{
Impute(X)
}
\arguments{
\item{X}{Is a numeric marker matrix}
}
\description{
Impute is a function to perfomr marker imputation and provde a list of monomorphic markers with missing values.
Monomorphic markers witout missing values have to be removed manually.
}
