% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NASAPOWER.R
\name{periodicextraction}
\alias{periodicextraction}
\title{Summarize climdata outputs}
\usage{
periodicextraction(
  df = NULL,
  locs.merged.climate = NULL,
  days = 120,
  last.day = 120
)
}
\arguments{
\item{df}{Is a one row (one location) data frame with the date of sowing (sowing.d), or a date from which the variables will start to be summarized.}

\item{locs.merged.climate}{Is a list of locations with the climate data}

\item{days}{Is the number of days after the initial date that will be consider to extract the data}

\item{last.day}{Is the number of days that will be used to average the data. i.e, las.day must be <= than days and multiple of days}
}
\description{
Function to summarize the climate date by groups of days for a single location. It requires the locs.merged.climate object which is the climate data.
Dates must be in the appropriate format otherwise errors are shown.
}
